Migrate from Angular 4.2 to Angular 5.0
I followed the instructions in [this conversion guide](https://angular-update-guide.firebaseapp.com)
and that seemed to have worked pretty well. Our project didn't use a lot
of the "super fancy" stuff that might have made it more difficult.

The one place where I had to change "content" was in `UserListService`,
where the uses of `Http` had to be changed to `HttpClient`, which also
changed how calls to the server were made.