Add tests for `UserListService`
I followed info on [the Angular guide to the `Http` class](https://angular.io/guide/http)
and [the Angular guide to testing](https://angular.io/guide/testing)
to write up some tests for the `UserListService` that mock the HTTP
requests in some quite nifty ways. We now have essentially 100% coverage
on this service. (There's 2/3 branch coverage, but this is apparently
something goofy about how TypeScript compiles imports, and there's
nothing to be done about it at the moment.)

Closes #45