Fix all the TSLint warnings
This fixes all the TSLint warnings, including a zillion places where
we'd used double quotes when single quotes are recommended.

Merging this in isn't likely to be fun if you've made any changes in
the `imports`, so be prepared for some merge conflict fun.

There are still a handful of IntelliJ warnings. Some are things I'm just
not sure what to do about (if anything); others are things I don't
think we'll do anything with. If we're going to ignore any, though,
we might want to document that in the write-up.