Move User filtering logic to service, improve filter efficiency
A couple issues are addressed in this commit. They were intertwined so I took care of them simultaneously.
1. Moved the filtering logic to user-list.service.ts from user-list.component.ts per #53.
2. Increased efficiency by only calling filterUsers() when one of the input values changes (instead of when *any* change is detected by Angular). The ngFor is bound now to an array rather than a function call. The array is updated when one of the filter input fields changes. This is per an issue in lab 4 (https://github.com/UMM-CSci-3601/3601-lab4_mongo/issues/29).

One further improvement that could be made here is to use Angular's reactive forms to access filter input values. Instead of binding the input event to a function call that takes `$event.target.value` as an argument, it would be more Angular-y to use a form group with named form controls.