Remove unneeded refreshUsers() from component
The only place this method was being called was in the init hook of the component, so I just moved the code directly into the init hook. I also removed the call to this.updateFilter(), since the filter doesn't need to be updated when the component is initialized (it was also causing the tests to fail).

Per #53