Update README.md to reflect current version of Angular in the lab
We were warning students of potential mismatches between the Angular in
the lab code and the most current version of Angular. This edit makes it
so that we just need to update what we say is used in this lab as opposed
to what is the most recent version. It acknowledges that there are
changes twice a year for sure and that we might not we might not match.
This commit also tones down the use of language that may be challenging
for non-native speakers of English (e.g., labeling something as 'insane').