Update LABTASKS.md
There are significant changes to the lab that require substantial changes to the LABTASKS. This commit addresses the need for some change. In addition to updating the tools and structures that are different, this commit also addresses more explicit changes about how to practice continuous integration with the testing practices and version control setup that we have.