Update LABTASKS
This is not totally complete. We still need to be sure these are the right questions to ask. We need to decide if the task of specifically looking at two things in Angular from that location is correct. I feel like maybe that's not the right place to point people.