Clean up mongoseed.sh
`mongoseed.sh` errors out in a couple of scenarios.

- It doesn't like it if there are no json files; in that case, it tries to literally import a file named `*.json` (which of course doesn't exist).
- It doesn't like it if any of your json files have spaces in their name, because `$file` isn't quoted.
- It doesn't like it if the path to mongoseed.sh has spaces, because `$BASH_SOURCE` isn't quoted. (But usually, this isn't an issue since in most invocations the path is just `.`)
- It has strange behavior if the $MONGO_DB environment variable contains spaces, because neither $MONGO_DB nor $SEED_DB are quoted (For example, if you do export MONGO_DB='dev --help' and then run mongoseed.sh, instead of seeding the database, it will give you the help pages for mongo and mongoimport.)

This commit fixes the first issue by checking to make sure that the file exists before importing it; this catches the case where `mongoseed.sh` tries to import a file named `*.json`. This commit fixes the three other issues by quoting variables in the appropriate places.

Also, this commit renames `$SEED_DB` to `$seed_db`. Since this is a local variable, not an environment variable, it gets a lower case name.