Update Cypress tests
These tests all seem to work and check for similar things that we checked for with Protractor. We probably want to remove comments that refer to the old way we did things, but I was not sure how people would feel about that or if the comments might help compare Protractor and Cypress more clearly while we are in this transition. In the end, I think we will want to remove those comments.

The biggest challenges:
- The use of 'check' for radio buttons wanted us to use a more boring radio button
- The use of 'select' wanted us to use a more basic selector (similar problem as check)
- Some of the Cypress tests will not work if the tests use `async`, and I didn't notice that right away, so I spent some time confused about that
- The tests that are working with `firstCompanyName` and `firstUserName` use a variable, and figuring out the scope of that variable so that we'd be able to use it after we change pages was just a little ugly