Remove commented out old approaches and update comments
This is the commit where someone probably wants to look if they want to see the old ways things were done. I'm removing them for readability, but there may be things people would want to look at to understand more about differences between Cypress and Protractor. 

This may have more comments than what are really needed, but I do think all of the remaining comments refer to code that is actually here and describe what it is actually doing.