Update almost all the Java libraries
This updates almost all the Java libraries to their latest versions.

The exception are the Jackson libraries, which are still behind, and in the case of `mongojack`, quite a ways behind. If I bring them up to date, things don't compile, and I'm not sure how to fix them. The examples in their GitHub repos simply don't appear to work.

So for now I'll leave them as they are, but we really should figure out how to fix them, or stop using them.