Suppress LGTM's weak crypto algorithm warning
LGTM is (appropriately) anxious about the use of MD5 since it's not cryptographically secure anymore. The use here is to contact Gravatar, which doesn't allow anything _other_ than MD5.

Doing some reading on this suggests that we really shouldn't be promoting the use of Gravatar as there are serious privacy concerns with it. (See, e.g., https://meta.stackexchange.com/questions/44717/is-gravatar-a-privacy-risk)

For now, though, I'll just suppress the warning and add an issue to remove the Gravatar code altogether as a separate task.