Add docs on updating badges and enabling PR checks
This removes the CodeScene and Codacy badges because I don't think they make sense in the class context.

I added a pseudo-badge at the top to remind folks to fix the badges and enable the PR checks, with a link to the new MD page that documents how to do that.