Close MongoClient in Server.java down if server fails or stops
There were warnings about the fact that our `MongoClient` isn't necessarily closed if something Bad happens and, for example, the server crashes. The recommendations were to use the try-with-resources or a `try` with a `finally` block, but those don't work here. The `MongoClient` variable has to be "live" the entire time that the server is up, so we can't shut it down anywhere in `main()`.

This adds server event handlers that close the client when either the server fails to actually start, or the server is stopping. I don't know if those are the right ones or enough, so feedback would certainly be appreciated.

I also don't know if this is several more lines of code and additional complication that doesn't actually make anything better, so thoughts there would be good as well.