Remove the name regex in AddUserComponent
It was a weird regex that didn't make much sense.

I could have tried to write a better one, but

1. personal names aren't really a good use-case for regular expressions—they don't have enough structure, and
2. we're demonstrating how to use regular expressions in plenty of other places, so there shouldn't be a big pedagogical impact.

I also removed one test that asserted names shouldn't have symbols in them, since it's hard to decide which exact symbols to allow. (cf. O'Neil, St. John, anyone with a hyphen)

I decided to leave the "should allow digits in the name" test, because that's still an assertion that holds. (And it's non-obvious that the name field would accept digits, so it makes sense to test for it.)