Clean up the regex on the age field in AddUserComponent
In addition to fixing the regex, I also removed the type="number" field from the <input> element.

type="number" was causing an issue where the input field wouldn't even recognize inputs that weren't formatted with a number. This was messing with our validation pipeline.

Unfortunately, now we don't get nice QoL features on this input like the spinner or the numeric keypad.

My reasoning for this change is that it's more important that our validation code works as advertised than it is to have a completely idiomatic form. 🤷🏻‍♂️ You all may disagree with that, though!

Oh, we still have an <input type="number"> on the main user page, so we're still showing it to the students in at least one place.