Use Document insted of BasicDBObject
Based on [this](https://stackoverflow.com/a/44879031) StackOverflow
answer, it looks like BasicDBObject was a precursor to Document, and that
Document does all the things BasicDBObject does, but more, and better.

BasicDBObject isn't deprecated or anything, but still, it's nice to be
consistent about which ones we use, and given the choice, I'd rather go
with the New Hotness.