Add a test that negative ages are rejected
Prof. McPhee noted that it wasn't immediately obvious that we're
handling negative ages correctly. (We are--they're caught by the
"age istoo low" validator--but it's good to have a test ensuring that.)

In the tests, I decided to enforce that negative numbers should get
caught in the "age is too low" validator, rather than in the regex
validator. This is because the "age is too low" validator gives a more
meaningful error messages.