Disable several Checkstyle checks
This disables several Checkstyle checks that are included in the Sun configuration. I like the spirit of these, but our experience is that these tend to confuse people more than help them. Also, the structure of the Java code in the project has become quite simple because the server library (e.g., Javalin) encapsulates the bulk of the design complexity, leaving us to just write code at the "leaves" of the design.

The disabled checks are:

   *  DesignForExtension: This only really makes sense if we're building a reasonably
         sophisticate OO library, and that kind of design really doesn't
         come up in this class, because the server library (e.g., Javalin)
         encapsulates all that design work, leaving us to just act around
         the "leaves" of the design.
   *  HideUtilityClassConstructor: If everything is static 
         in a class, then it shouldn't have
         an accessible constructor. This is both subtle and
         confusing, and not likely to come up much in the class,
         so we'll just leave it out.
   * FinalParameters: I really likes the *intent* of this rule,
         but it's really trying to use CheckStyle to fix
         a mistake in the design of Java. This just confuses
         students, and requires a ton of extra typing
         that probably doesn't buy folks much.