Replace all HTTP response numbers with named constants in Java code
StackOverflow seems to think that there is no standard Java library class that contains *all* the standard HTTP codes, but `HttpURLConnection` seems to have a really substantial collection, and probably anything we're likely to use. So I imported that and used it's names (HTTP_OK and HTTP_CREATED) in place of 200 and 201.

We should probably try to do something similar for the JavaScript/TypeScript code.