Added configuration info for Checkstyle VS Code plugin
This uses the "rianadon.vscode-java-checkstyle" Checkstyle plugin, but that doesn't seem to be a super robust option. The one with the most downloads ("shengchen.vscode-checkstyle") has pretty poor reviews, and I couldn't actually get it to work. It would be nice to have VSCode support, but maybe it's not absolutely vital?