Regenerate package-lock.json
By deleting it and then running `npm install`. This should update many
of the transitive dependencies.

It was necessary to change dependencies on `jasmine-core` and `ts-node`
to allow a more inclusive range of versions. (Some other packages wanted
peer dependencies on higher versions of `jasmine-core` and `ts-node`, so
we had to allow that.)