Update to mongojack 4.x.x
Including fixing a new change in the API related to how UUIDs are
stored.

Previously, the Mongo driver for Java had a weird, non-standard way of
storing UUID values in the database. Nowadays, you can either use the
"standard" way of encoding UUIDs, or the "legacy java" way of encoding
UUIDs (depending on whether you still have legacy Java data in your
database.)

We don't have any databases with legacy UUID data, so we can just use
the standard encoding and we should be fine.