Migrate code from Javalin 3 to 4
I think this may make all the necessary changes to move our code from Javalin 3 to 4.

This took a lot longer than it did on, e.g., Lab 2, mostly because there were a lot of issues with parsing JSON strings to objects that confused me for a while.

My Mongo install in my office is messed up, so I can't actually run the system. I'll push this and we'll see what GitHub Actions does, and then I'll run it more carefully in the lab or at home where Mongo works.

I should also see if this has any impact on the lab write-up.