Fix a race condition in cypress/.../app.spec.ts
We were clicking the "open sidenav" button, and then expecting the
"open sidenav" button to be visible. This will *only* be the case for a
split-second before the sidenav appears.

I would guess that this has always been an issue, but it's started
failing consistently on GitHub actions as of Cypress version 9.