Replace one use of JS string `+` with template literal
I'm just using VS Code's global search tool to look for +s in JS/TS code, and then convert them to use template literals. This is my first one!

The Karma tests didn't fail if I broke this, but the E2E tests did, which was nice since I forgot to switch to backticks at first and actually broke things. The tests caught it, though, and I think this is good now.