Revert the Cypress refactorings
I had attempted to refactor some of the Cypress E2E tests in an attempt to make BetterCodeHub less grumpy, but everyone agreed (on the PR) that this had not actually helped anything. This undoes those refactorings as a single commit. I may try to do some squash action on all this when we merge in the PR to reduce some of the noise here.