Exclude E2E tests from BetterCodeHub checks
This tells BetterCodeHub to ignore the Cypress E2E tests.

I think part of the problem with BetterCodeHub was the change from `/client/e2e` to `/client/cypress`. That change wasn't reflected in the BetterCodeHub configuration, which is why I think it started fussing about the E2E tests all of a suddent.