Add Thunder Client as a recommended extension
This commit adds information about Thunder Client to the `RESOURCES.md` file and `DEVELOPMENT.md` file as well as actually putting Thunder Client in the `extensions.json` file so that it will show up as being a recommended VS Code extension. There is a link back to Lab 2 where Thunder Client is first introduced to the students (with a few screen shots).