Add reference to `server side` in contrast to `client side`
The documentation described in `DEVELOPMENT.md` referred to the idea of `ng serve` serving up the _client side_ of the project, but never mentioned that the `./gradlew run` is running or serving up the _server side_ of the application. Since the language there is potentially confusing (since we "serve" up the "client side" of things, I thought it makes sense to give a name to the back end... the _server side_ of things in the client/server architecture.