Update style information for radio buttons
It looks like maybe the style information was for something with mdc where we don't actually have the mdc version of things yet. I removed the mdc part and altered the style a little bit. I'm hoping this will allow the Cypress tests to find the button now. The tests pass locally for me with this minor change.