Add some validation for age and role
Since some of this is validated when something is created, I thought it might make sense to also make sure that the queries are valid here. This kind of validation could be done on the client side. I'm honestly not sure where it would work best. But, this can be used to give us more error messages that are specific to the problem if a query is somehow entered that is not valid.