Explain away the magicness of 150
This number isn't all that magical. By naming it according to its intent/reason to exist, the checkstyle doesn't think we are using it in a magical way. This validation seems like a good idea since it prevents people from entering unreasonably large numbers for age of users, which might indicate an error in data entry for creating or searching for users, so I think it's worth including.