Add several tests
I finally figured out how to test queries that included some validation of the parameters. I still need to figure out some more about how to throw errors and exceptions the "right" way in Javalin, and how to see the messages from those errors. Thunder Client may be helpful here, but I don't yet know how to access the body of the JSON that is returned for various problems with validation.