Add test for addUser
This one was a bit tricky for me, and I got some help on the Javalin Discord to figure out how to make an appropriate call to`bodyValidator`. I also changed the UserController a tiny bit by moving the status update to be the last thing that happens. I want to be absolutely sure that we have done all the other things before we claim to have successfully created the user.