Finish server tests
Coverage is enough now so that this passes the build checks. There is a branch about sorting in descending order that has not been touched, and the avatar generation (a private method) is not tested for cases when it doesn't work. Now we just need to decide if we want to look at some functional tests or not for future semesters. I don't know if I'll do that this semester or not.