Attempt to fix e2e tests by watching chained events
My understanding is that `.should` will wait until the should is met. By altering the length to be specific about how many cards we should see, I think it might be less flaky. The test may be a bit more brittle this way since it relies on knowing how many cards, but the alternative (if this helps) doesn't seem better.