Choose more informative documentation (given the option)
Comparing the iteration template to the labs, the iteration template is more likely to be used as a quick, definite reference by students so that it's all in one place. Since it can always be deleted, I think we should choose the more informative documentation so that students can have an easier time understanding the most likely point of reference.