Change user-list component's use of subscription
The way we were using subscribe was deprecated. Its signature involved two inputs to the subscribe, one of which was the error, and that signature is not recommended or supported anymore. Instead, this uses a structure that tells what to do when the `next` change is emitted from the observable, what to do for `error`, and what to do when changes to the observable are `complete`. 

I also added some comments to help describe what's happening here since observables and asynchronous behaviors are new to most students in the course, and it's tough to understand or know where to look for more description about it.