Change page object chaining and remove comment about fragility
I had a previous concern that the chaining in the Cypress test might be fragile, so I commented as such. It was suggested that the comment should be removed, so instead of leaving the comment I attempted to alleviate the chaining problem. This uses cy.get on the options that show up, which should remove the fragility from this chaining.