Undo changes to user-profile.component.ts
In commit a363fd25c2175b958202390dea1003fc20dbc41c I removed some code that should not be removed. Eventually, we will address the changes that are needed for how to subscribe and unsubscribe using pipe, map, and other things, but for now this will go back to how it was.