Remove use of `console` from code that will execute in the browser
Deep Source suggested that we remove this use of `console`.

It is considered a best practice to avoid the use of any `console` methods in JavaScript code that will run on the browser. Usually, `console` methods are only used for debugging, and can leak internal info to the client. Removing the `console` call will fix this issue.