Improve error handling
Deep Source suggested that we should use `e` since it was declared but not used. I found a source that said a bit more about error handling that seemed to suggest we could use the error rather than just sending the same message each time.

I also added handling for client errors, but I didn't add testing for either of these yet. I'm not sure how to generate these problems for unit testing to see that the correct messages are presented.

I also removed the message to the `console`.

https://www.positronx.io/angular-error-handling-tutorial-with-examples/#:~:text=The%20simplest%20way%20to%20handle,object%20with%20error%20status%20code.