Make instance methods use `this`
Deep Source suggests that if a class method does not use this, it can sometimes be made into a static function. If you do convert the method into a static function, instances of the class that call that particular method have to be converted to a static call as well (MyClass.callStaticMethod())

Since these are specifically page objects, I think it's better to use `this` than to make things static. One benefit I'm seeing is that all of the selectors that might change are at the top now. I think other things within the methods will be stable.