Use `providedIn` property
The `providedIn` property is mandatory for `Injectables`. Using the `providedIn` property makes classes decorated with `@Injectable` tree shakeable.

If you use `providedIn`, the injectable is registered as a provider of the module without adding it to the provider list of the module.

This is suggested by Deep Source, and I think it seems less likely to end up with problems forgetting to edit the list of providers in the module.