Make optional value match Lab 4
There was not a particular reason listed about why this was changed, and this matches the Lab. I think this is more readable even though the type assignment is not needed since it's clearly a boolean. Perhaps some other source will tell me a good reason to change it back.