Remove checks of users returned from `getUsers`
Since `getUsers` is a server function, and this is a test of the client side user service, we couldn't meaningfully test (or care about) what gets returned from those server-side filters.

We might want to update our spy to not even return the users since it was all for show anyhow, but we might consider leaving that in so students can see what it's like.