Add foojay Disco Gradle plugin to download JDKs
Our intent in a previous Gradle update had been to set things up so that Gradle would automatically download required JDKs if they weren't installed/available. That never actually worked.

Returning to it, it seems like (now?) you need to specify a resolver plugin for that to work. This adds the "foojay Disco Installer" (seriously?) which does seem to work. It downloaded a JDK 11 on my laptop which then seemed to work.