Add tests for `AddUserComponent#submitForm`
This adds tests for both the success and error paths for `AddUserComponent#submitForm`.

This gets the test coverage for `AddUserComponent` to 100%, which is pretty nifty.

The initial version of this code was suggested by GitHub CoPilot, but I had to make several changes to get it to compile and pass. I also added a variety of comments that hopefully helps explain what's going on.