Fix problem with `UserService#addUser` test
There were some notes about the return value for the mocked call to `httpClient` not working. The problem was that the tests were assuming that it was returning a string, when in fact it returns a JSON object with an `:id` field that has the string as its value.

This fixes all that and gets the tests to pass.

Closes #1307