Refactor `Server` class
This is a collection of mostly minor refactorings to the `Server` class, mostly based on suggestions by GitHub CoPilot.

Some are really just about changing the order of things, where other are more substantive. One, for example, is injecting the `mongoClient` and `userController` into the constructor for `Server`. I think this will make it easier to test things.