Create `Controller` interface
This reduces the coupling between the Server and the controllers, with the only reference to a concrete instance of a `Controller` being in the call to the `UserController` constructor in `main()`.

Once the controllers are constructed and passed to the `Server` constructor, the `Server` knows nothing about them other than that they provide an `addRoutes()` method, which is called in `Server#setupRoutes()`.