Create `Main.java`
This moves the `main()` method and a couple of helper methods to the new `Main` class. The moved methods include the untestable `getEnvOrDefault()` method and the `getControllers` method, which is tied to specific controllers.

The nice thing here is that (along with the introduction of the `Controller` interface), the `Server` class now has _no_ dependencies on any specific controllers. The `Main#getControllers` method encapsulates all the dependencies on specific controllers, and `Server` only knows about the `Controller` interface.