Rename the `Controller` interface to `AddsRoutes`
@willster000 pointed out that the documentation on the `Controller` interface was potentially confusing, so I expanded it.

In the process, I actually renamed the `Controller` interface to be the `AddsRoutes` interface since that is really what this interface is about. It doesn't really describe whole controllers, it just describes this specific aspect of controllers.

While I think this is in some ways the "right" thing to do, it's possible it might make things more confusing for the students given their limited experience with interfaces.