Removed `any` from the `MATERIAL_MODULES` declaration
DeepSource was grumpy because the automated switch to standalone components introduced a use of `any`.

The problem is that `MATERIAL_MODULES` is an array of types without any super class or obvious way to type the elements of the array. So I just removed the type declaration altogether.

I also organized the inputs while I was there, since they were kind of a mess.