Improve error handling in `UserProfile` when user fails to load
Previously when I user failed to load, we got no useful information like the HTTP status from the server, and there was no navigation away from the failed page.

This updates the display to:

- Display the user if it exists
- Display the error message if the user doesn't exist but the error does
- Display a loading message otherwise

We added an `error` field in `UserProfile` that contains a `help` and `message` field, which we use to display a hopefully helpful note along with the actual error message we received from the server.

We also added a button that allows the user to navigate back to the user list when there's an error.

We added a `error-card` CSS class with a red background to highlight the error message.