Modify display of user card fields
This removes the headers and just makes them `<p>` with `matListItemTitle` doing the formatting to make the titles stand out.

We had to make the role and age entries all on the same line to avoid an annoying leading space. Still not sure why we had to do that, but it fix the problem.