Replace long string appends with multi-line strings
This replaced a bunch of unreadable and unmaintainable string append messes with single multi-line strings. These are much more readable and maintainable.

GitHub CoPilot was able to do most of this automatically, although I had to run it on each string separately. I suspect there's a prompt that would have had it change all of them in the file, but I didn't find it.