Update when client and server tests are checked
For some reason, the client and server tests aren't being run on the PR that @JustusFluegel created. I'm not sure why, but I'm wondering if we just don't have the requirements properly configured in the workflow files. This updates them to match the CodeQL settings (which do seem to work) to see if that fixes things.