Broaden workflow triggers
I think that the simple `on: [push]` that I used doesn't work if there is a pull request from a fork. This changes all the client and server test workflows to use the same configuration as the CodeQL configuration in the hopes that this gets things to work for us.