"Modernize" the ESLint config
What we had contained some now quite outdated connections to TSLint, which hasn't been a thing for several years.

Based on the discussion in this short video:

https://github.com/angular-eslint/angular-eslint/issues/1388#issuecomment-1566153131

I basically grabbed the "default" config file that he shows in the video, and moved a couple of things over from our old config file (like ignoring the fact that `_id` doesn't have the right form).

This created a few new warnings that I now have to deal with.