Convert `UserCardComponent` to use input signals
This is the result of using `ng g ngxtension:convert-signal-inputs` to try to convert our project to use input signals. We only have one component that uses `@Input`, namely `UserCardComponent`, so it's the only one that is changed by this migration.

The migration provided a really general "solution" with things like `<User | undefined>`, so I introduced `.required()` to clean some of that up. I also had to add `input()` to the `user-card.component.spec.ts` file since `component.user` is now an input signal instead of "just" a `User`.