Fix broken Karma tests by moving `zone.js` import to top in `test.ts`
The `zone.js` import statement in `test.ts` *must* be first, before any other imports. It actually patches the async APIs in the browser to make them "test friendly". If we import it after other modules, those modules won't "see" the patched APIs and will continue to use the native browser async APIs, which will cause tests to fail.

This makes the blank link after this `import` important; without it running "Organize Imports" in VS Code will move the `zone.js` import to the end of the imports, which will cause tests to fail.