Get the basic communication to work
I was throwing errors because I was deleting contexts from my `HashSet` while I was iterating over it, and that's not allowed.

My return type is a mess, and needs to be fixed.

My clients are also disconnecting really quickly, so I need to figure out how to fix that.